Update patch set 1
Patch Set 1: Code-Review-1

If you want to run Python tests, there is no need to use a Dockerfile.  Add a tox.ini file, if you have not done so already, and we can configure a Python job to do the work.  I think that will be much easier.  The LF supports Python 3.6 and 3.7.

Patch-set: 1
Label: Code-Review=-1